fix ambient light plugging

+++ includes/omnilight.h
@@ -6,7 +6,7 @@
 /*   By: darbib <darbib@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/25 16:26:46 by darbib            #+#    #+#             */
-/*   Updated: 2020/05/27 18:03:47 by darbib           ###   ########.fr       */
+/*   Updated: 2020/07/10 22:36:26 by darbib           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ typedef struct		s_olight
 {
 	t_vect			pos;
 	t_vect			current_pos;
-	float			ratio;
+	double			ratio;
 	unsigned char	rgb[3];
 }					t_olight;
